Upgrade to cURL 7.19.4, which fixes CVE-2009-0037. Omitted from the BZFlag subversion repository are the "packages" and "tests" subdirectories because they are large and beyond our scope. Upstream files that are created during our build process (e.g., Makefile.in, Makefile, *.pdf, *.html) are also omitted.
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@19517 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/other/curl/docs/examples/simplepost.c
@@ -5,7 +5,7 @@
  *                            | (__| |_| |  _ <| |___
  *                             \___|\___/|_| \_\_____|
  *
- * $Id: simplepost.c,v 1.3 2007-07-12 21:11:10 danf Exp $
+ * $Id: simplepost.c,v 1.4 2008-05-22 21:20:09 danf Exp $
  */
 
 #include <stdio.h>
@@ -26,7 +26,7 @@ int main(void)
 
     /* if we don't provide POSTFIELDSIZE, libcurl will strlen() by
        itself */
-    curl_easy_setopt(curl, CURLOPT_POSTFIELDSIZE, strlen(postthis));
+    curl_easy_setopt(curl, CURLOPT_POSTFIELDSIZE, (long)strlen(postthis));
 
     res = curl_easy_perform(curl);
